+++ SharpSploit/Execution/DynamicInvoke/Generic.cs
@@ -24,17 +24,9 @@ namespace SharpSploit.Execution.DynamicInvoke
         /// <returns>Object returned by the function. Must be unmarshalled by the caller.</returns>
         public static object DynamicAPIInvoke(string DLLName, string FunctionName, Type FunctionDelegateType, ref object[] Parameters)
         {
-            IntPtr hModule = Execution.Win32.Kernel32.LoadLibrary(DLLName);
-
-            IntPtr pFunction = Execution.Win32.Kernel32.GetProcAddress(hModule, FunctionName);
-
-            Delegate funcDelegate = Marshal.GetDelegateForFunctionPointer(pFunction, FunctionDelegateType);
-
-            Object result = funcDelegate.DynamicInvoke(Parameters);
-
-            return result;
+            IntPtr pFunction = GetLibraryAddress(DLLName, FunctionName);
+            return DynamicFunctionInvoke(pFunction, FunctionDelegateType, ref Parameters);
         }
-
 
         /// <summary>
         /// Dynamically invokes an arbitrary function from a pointer. Useful for manually mapped modules or loading/invoking unmanaged code from memory.
@@ -47,10 +39,7 @@ namespace SharpSploit.Execution.DynamicInvoke
         public static object DynamicFunctionInvoke(IntPtr FunctionPointer, Type FunctionDelegateType, ref object[] Parameters)
         {
             Delegate funcDelegate = Marshal.GetDelegateForFunctionPointer(FunctionPointer, FunctionDelegateType);
-
-            Object result = funcDelegate.DynamicInvoke(Parameters);
-
-            return result;
+            return funcDelegate.DynamicInvoke(Parameters);
         }
 
         /// <summary>